feat: add all CRUD methods, service and exception classes

+++ src/main/kotlin/ru/vladamamutova/repository/PersonRepository.kt
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository
 import ru.vladamamutova.domain.Person
 
 @Repository
-interface PersonRepository : CrudRepository<Person, Long> {
+interface PersonRepository : CrudRepository<Person, Int> {
 }
